Ask for Sync

+++ Trackcat/app/src/main/java/de/trackcat/Database/Models/User.java
@@ -350,7 +350,7 @@ public class User {
      *
      * @return value of type long
      */
-    public float getTimeStamp() {
+    public long getTimeStamp() {
         return timeStamp;
     }
